ocr/dev-requirements.txt
pytesseract
Pillow
redis
